Graphics.Rendering.OpenGL.GL.BufferMode

Description

This is a purely internal module for (un-)marshaling BufferMode.

Documentation

data BufferMode

The set of color buffers which are selected for reading and writing.

Constructors

NoBuffers
No color buffers are selected.

FrontLeftBuffer
Only the front left color buffer is selected.

FrontRightBuffer
Only the front right color buffer is selected.

BackLeftBuffer
Only the back left color buffer is selected.

BackRightBuffer
Only the back right color buffer is selected.

FrontBuffers
Only the front left and front right color buffers are selected. If there is no front right color buffer, only the front left color buffer is selected.

BackBuffers
Only the back left and back right color buffers are selected. If there is no back right color buffer, only the back left color buffer is selected.

LeftBuffers
Only the front left and back left color buffers are selected. If there is no back left color buffer, only the front left color buffer is selected.

RightBuffers
Only the front right and back right color buffers are selected. If there is no back right color buffer, only the front right color buffer is selected.

FrontAndBackBuffers
All the front and back color buffers (front left, front right, back left, back right) are selected. If there are no back color buffers, only the front left and front right color buffers are selected. If there are no right color buffers, only the front left and back left color buffers are selected. If there are no right or back color buffers, only the front left color buffer is selected.

AuxBuffer GLsizei
Only the given auxiliary color buffer no. i is selected.

FBOColorAttachment GLsizei
Only the given color attachment of the bound framebufferobject is selected for reading or writing.

marshalBufferMode :: BufferMode -> Maybe GLenum

unmarshalBufferMode :: GLenum -> BufferMode

unmarshalBufferModeSafe :: GLenum -> Maybe BufferMode

maxColorAttachments :: GLenum