Graphics.Rendering.OpenGL.GL.EdgeFlag

Description

This is a purely internal module for (un-)marshaling EdgeFlag.

Documentation

data EdgeFlag

A vertex can begin an edge which lies in the interior of its polygon or on the polygon's boundary.

Constructors

BeginsInteriorEdge

BeginsBoundaryEdge

marshalEdgeFlag :: EdgeFlag -> GLboolean

unmarshalEdgeFlag :: GLboolean -> EdgeFlag