Graphics.Rendering.OpenGL.GL.FramebufferObjects.Queries

Description

Documentation

data AttachmentObjectType

Constructors

DefaultFramebufferAttachment

TextureAttachment

RenderbufferAttachment

attachmentObjectType :: FramebufferAttachment fba => FramebufferTarget -> fba -> GettableStateVar (Maybe AttachmentObjectType)

attachmentObject :: FramebufferAttachment fba => FramebufferTarget -> fba -> GettableStateVar (Maybe (Either RenderbufferObject TextureObject))

tries to retrieve the object that is bound to the attachment point of the given framebuffertarget. If the object type of it is None or the default, then Nothing is returned, otherwise the bound RenderbufferObject or TextureObject

attachmentTextureLayer :: FramebufferAttachment fba => FramebufferTarget -> fba -> GettableStateVar GLint

attachmentTextureLevel :: FramebufferAttachment fba => FramebufferTarget -> fba -> GettableStateVar Level

attachmentTextureTextureTargetCubeMapFace :: FramebufferAttachment fba => FramebufferTarget -> fba -> GettableStateVar TextureTargetCubeMapFace

attachmentRedSize :: FramebufferAttachment fba => FramebufferTarget -> fba -> GettableStateVar GLint

attachmentBlueSize :: FramebufferAttachment fba => FramebufferTarget -> fba -> GettableStateVar GLint

attachmentGreenSize :: FramebufferAttachment fba => FramebufferTarget -> fba -> GettableStateVar GLint

attachmentAlphaSize :: FramebufferAttachment fba => FramebufferTarget -> fba -> GettableStateVar GLint

attachmentDepthSize :: FramebufferAttachment fba => FramebufferTarget -> fba -> GettableStateVar GLint

attachmentStencilSize :: FramebufferAttachment fba => FramebufferTarget -> fba -> GettableStateVar GLint

renderbufferWidth :: RenderbufferTarget -> GettableStateVar GLsizei

renderbufferHeight :: RenderbufferTarget -> GettableStateVar GLsizei

renderbufferInternalFormat :: RenderbufferTarget -> GettableStateVar PixelInternalFormat

renderbufferSamples :: RenderbufferTarget -> GettableStateVar Samples

renderbufferRedSize :: RenderbufferTarget -> GettableStateVar GLint

renderbufferBlueSize :: RenderbufferTarget -> GettableStateVar GLint

renderbufferGreenSize :: RenderbufferTarget -> GettableStateVar GLint

renderbufferAlphaSize :: RenderbufferTarget -> GettableStateVar GLint

renderbufferDepthSize :: RenderbufferTarget -> GettableStateVar GLint

renderbufferStencilSize :: RenderbufferTarget -> GettableStateVar GLint