Graphics.Rendering.OpenGL.GL.VertexSpec

Description

This module corresponds to section 2.7 (Vertex Specification) of the OpenGL 2.1 specs.

Vertex Coordinates

class Vertex a where

Specify the (x, y, z, w) coordinates of a four-dimensional vertex. This must only be done during renderPrimitive, otherwise the behaviour is unspecified. The current values of the auxiliary vertex attributes are associated with the vertex.

Note that there is no such thing as a "current vertex" which could be retrieved.

Methods

vertex :: a -> IO ()

vertexv :: Ptr a -> IO ()

class VertexComponent a

The class of all types which can be used as a vertex coordinate.

Minimal complete definition

vertex2, vertex3, vertex4, vertex2v, vertex3v, vertex4v

Auxiliary Vertex Attributes

Apart from its coordinates in four-dimensional space, every vertex has associated auxiliary attributes: Its texture coordinates, a normal, a fog coordinate, and a color plus a secondary color. For every attribute, the OpenGL state contains its current value, which can be changed at any time.

Every attribute has a "natural" format via which it can be manipulated directly as part of the OpenGL state, e.g. the current texture coordinates are internally handled as TexCoord4 GLfloat. Different formats are converted to this format, e.g. the s, r, and t coordinates of a TexCoord3 GLint are converted to floating point values and a q coordinate of 1.0 is implicitly assumed.

Consequently, the vast majority of classes, functions, and data types in this module are for convenience only and offer no additional functionality.

Texture Coordinates

currentTextureCoords :: StateVar (TexCoord4 GLfloat)

The current texture coordinates (s, t, r, q) for the current texture unit (see activeTexture). The initial value is (0,0,0,1) for all texture units.

class TexCoord a where

Change the current texture coordinates of the current or given texture unit.

Methods

texCoord :: a -> IO ()

texCoordv :: Ptr a -> IO ()

multiTexCoord :: TextureUnit -> a -> IO ()

multiTexCoordv :: TextureUnit -> Ptr a -> IO ()

class TexCoordComponent a

The class of all types which can be used as a texture coordinate.

Minimal complete definition

texCoord1, texCoord2, texCoord3, texCoord4, texCoord1v, texCoord2v, texCoord3v, texCoord4v, multiTexCoord1, multiTexCoord2, multiTexCoord3, multiTexCoord4, multiTexCoord1v, multiTexCoord2v, multiTexCoord3v, multiTexCoord4v

newtype TexCoord1 a

Texture coordinates with t=0, r=0, and q=1.

Constructors

TexCoord1 a

data TexCoord2 a

Texture coordinates with r=0 and q=1.

Constructors

TexCoord2 !a !a

data TexCoord3 a

Texture coordinates with q=1.

Constructors

TexCoord3 !a !a !a

data TexCoord4 a

Fully-fledged four-dimensional texture coordinates.

Constructors

TexCoord4 !a !a !a !a

Normal

currentNormal :: StateVar (Normal3 GLfloat)

The current normal (x, y, z). The initial value is the unit vector (0, 0, 1).

class Normal a where

Change the current normal. Integral arguments are converted to floating-point with a linear mapping that maps the most positive representable integer value to 1.0, and the most negative representable integer value to -1.0.

Normals specified with normal or normalv need not have unit length. If normalize is enabled, then normals of any length specified with normal or normalv are normalized after transformation. If rescaleNormal is enabled, normals are scaled by a scaling factor derived from the modelview matrix. rescaleNormal requires that the originally specified normals were of unit length, and that the modelview matrix contains only uniform scales for proper results. Normalization is initially disabled.

Methods

normal :: a -> IO ()

normalv :: Ptr a -> IO ()

class NormalComponent a

The class of all types which can be used as a component of a normal.

Minimal complete definition

normal3, normal3v

data Normal3 a

Constructors

Normal3 !a !a !a

Fog Coordinate

currentFogCoord :: StateVar (FogCoord1 GLfloat)

The current fog coordinate. The initial value is 0.

class FogCoord a where

Change the current fog coordinate.

Methods

fogCoord :: a -> IO ()

fogCoordv :: Ptr a -> IO ()

class FogCoordComponent a

The class of all types which can be used as the fog coordinate.

Minimal complete definition

fogCoord1, fogCoord1v

newtype FogCoord1 a

A fog coordinate.

Constructors

FogCoord1 a

Color and Secondary Color

rgbaMode :: GettableStateVar Bool

If rgbaMode contains True, the color buffers store RGBA value. If color indexes are stored, it contains False.

currentColor :: StateVar (Color4 GLfloat)

class Color a where

Change the current color.

Methods

color :: a -> IO ()

colorv :: Ptr a -> IO ()

currentSecondaryColor :: StateVar (Color3 GLfloat)

class SecondaryColor a where

Change the current secondary color.

Methods

secondaryColor :: a -> IO ()

secondaryColorv :: Ptr a -> IO ()

class ColorComponent a

The class of all types which can be used as a color component.

Minimal complete definition

color3, color4, color3v, color4v, secondaryColor3, secondaryColor3v

data Color3 a

Constructors

Color3 !a !a !a

data Color4 a

A fully-fledged RGBA color.

Constructors

Color4 !a !a !a !a

currentIndex :: StateVar (Index1 GLint)

class Index a where

Change the current color index.

Methods

index :: a -> IO ()

indexv :: Ptr a -> IO ()

class IndexComponent a

The class of all types which can be used as a color index.

Minimal complete definition

index1, index1v

newtype Index1 a

A color index.

Constructors

Index1 a

Generic Vertex Attributes

data IntegerHandling

Constructors

ToFloat

ToNormalizedFloat

KeepIntegral

newtype AttribLocation

Constructors

AttribLocation GLuint

currentVertexAttrib :: AttribLocation -> StateVar (Vertex4 GLfloat)

currentVertexAttribI :: AttribLocation -> StateVar (Vertex4 GLint)

currentVertexAttribIu :: AttribLocation -> StateVar (Vertex4 GLuint)

class VertexAttrib a where

Methods

vertexAttrib :: IntegerHandling -> AttribLocation -> a -> IO ()

vertexAttribv :: IntegerHandling -> AttribLocation -> Ptr a -> IO ()

class (Storable a, Num a) => VertexAttribComponent a where

The class of all types which can be used as a generic vertex attribute. NOTE: Do not use the methods of this class directly, they were only exported by accident and will be hidden in future versions of this package.

Minimal complete definition

vertexAttrib4v, vertexAttrib4Nv, vertexAttrib4Iv

Methods

vertexAttrib1 :: AttribLocation -> a -> IO ()

vertexAttrib2 :: AttribLocation -> a -> a -> IO ()

vertexAttrib3 :: AttribLocation -> a -> a -> a -> IO ()

vertexAttrib4 :: AttribLocation -> a -> a -> a -> a -> IO ()

vertexAttrib1N :: AttribLocation -> a -> IO ()

vertexAttrib2N :: AttribLocation -> a -> a -> IO ()

vertexAttrib3N :: AttribLocation -> a -> a -> a -> IO ()

vertexAttrib4N :: AttribLocation -> a -> a -> a -> a -> IO ()

vertexAttrib1I :: AttribLocation -> a -> IO ()

vertexAttrib2I :: AttribLocation -> a -> a -> IO ()

vertexAttrib3I :: AttribLocation -> a -> a -> a -> IO ()

vertexAttrib4I :: AttribLocation -> a -> a -> a -> a -> IO ()

vertexAttrib1v :: AttribLocation -> Ptr a -> IO ()

vertexAttrib2v :: AttribLocation -> Ptr a -> IO ()

vertexAttrib3v :: AttribLocation -> Ptr a -> IO ()

vertexAttrib4v :: AttribLocation -> Ptr a -> IO ()

vertexAttrib1Nv :: AttribLocation -> Ptr a -> IO ()

vertexAttrib2Nv :: AttribLocation -> Ptr a -> IO ()

vertexAttrib3Nv :: AttribLocation -> Ptr a -> IO ()

vertexAttrib4Nv :: AttribLocation -> Ptr a -> IO ()

vertexAttrib1Iv :: AttribLocation -> Ptr a -> IO ()

vertexAttrib2Iv :: AttribLocation -> Ptr a -> IO ()

vertexAttrib3Iv :: AttribLocation -> Ptr a -> IO ()

vertexAttrib4Iv :: AttribLocation -> Ptr a -> IO ()

Texture Units

newtype TextureUnit

Identifies a texture unit via its number, which must be in the range of (0 .. maxTextureUnit).

Constructors

TextureUnit GLuint

maxTextureUnit :: GettableStateVar TextureUnit

An implementation must support at least 2 texture units, but it may support up to 32 ones. This state variable can be used to query the actual implementation limit.