Graphics.Rendering.OpenGL.GLU.ErrorsInternal

Description

This is a purely internal module corresponding to some parts of section 2.5 (GL Errors) of the OpenGL 2.1 specs and chapter 8 (Errors) of the GLU specs.

Documentation

data Error

GL/GLU errors consist of a general error category and a description of what went wrong.

Constructors

Error ErrorCategory String

data ErrorCategory

General GL/GLU error categories

Constructors

InvalidEnum

InvalidValue

InvalidOperation

InvalidFramebufferOperation

OutOfMemory

StackOverflow

StackUnderflow

TableTooLarge

TesselatorError

NURBSError

getErrors :: IO [Error]

recordErrorCode :: GLenum -> IO ()

recordInvalidEnum :: IO ()

recordInvalidValue :: IO ()

recordOutOfMemory :: IO ()