Graphics.Rendering.OpenGL.GL.BeginEnd

Description

This module corresponds to section 10.8 (Drawing Commands Using Begin and End) of the OpenGL 4.4 specs.

Begin and End Objects

data PrimitiveMode

Specification of the way the vertices given during renderPrimitive are interpreted. In the description of the constructors, n is an integer count starting at one, and N is the total number of vertices specified.

Constructors

Points
Treats each vertex as a single point. Vertex n defines point n. N points are drawn.

Lines
Treats each pair of vertices as an independent line segment. Vertices 2n-1 and 2n define line n. N/2 lines are drawn.

LineLoop
Draws a connected group of line segments from the first vertex to the last, then back to the first. Vertices n and n+1 define line n. The last line, however, is defined by vertices N and 1. N lines are drawn.

LineStrip
Draws a connected group of line segments from the first vertex to the last. Vertices n and n+1 define line n. N-1 lines are drawn.

Triangles
Treats each triplet of vertices as an independent triangle. Vertices 3n-2, 3n-1, and 3n define triangle n. N/3 triangles are drawn.

TriangleStrip
Draws a connected group of triangles. One triangle is defined for each vertex presented after the first two vertices. For odd n, vertices n, n+1, and n+2 define triangle n. For even n, vertices n+1, n, and n+2 define triangle n. N-2 triangles are drawn.

TriangleFan
Draws a connected group of triangles. One triangle is defined for each vertex presented after the first two vertices. Vertices 1, n+1, and n+2 define triangle n. N-2 triangles are drawn.

Quads
Treats each group of four vertices as an independent quadrilateral. Vertices 4n-3, 4n-2, 4n-1, and 4n define quadrilateral n. N/4 quadrilaterals are drawn.

QuadStrip
Draws a connected group of quadrilaterals. One quadrilateral is defined for each pair of vertices presented after the first pair. Vertices 2n-1, 2n, 2n+2, and 2n+1 define quadrilateral n. N/2-1 quadrilaterals are drawn. Note that the order in which vertices are used to construct a quadrilateral from strip data is different from that used with independent data.

Polygon
Draws a single, convex polygon. Vertices 1 through N define this polygon.

renderPrimitive :: PrimitiveMode -> IO a -> IO a

Delimit the vertices that define a primitive or a group of like primitives.

Only a subset of GL commands can be used in the delimited action: Those for specifying vertex coordinates (vertex, vertexv), vertex colors (color, colorv, secondaryColor, secondaryColorv, index, indexv), normal (normal, normalv), texture coordinates (texCoord, texCoordv, multiTexCoord, multiTexCoordv), and fog coordinates (fogCoord, fogCoordv). Additionally, evalPoint1, evalPoint2, evalCoord1, evalCoord1v, evalCoord2, evalCoord2v, materialAmbient, materialDiffuse, materialAmbientAndDiffuse, materialSpecular, materialEmission, materialShininess, callList, callLists, and setting edgeFlag are allowed. Writing the respective state variables is allowed in the delimited action, too.

Regardless of the chosen PrimitiveMode, there is no limit to the number of vertices that can be defined during a single renderPrimitive. Lines, triangles, quadrilaterals, and polygons that are incompletely specified are not drawn. Incomplete specification results when either too few vertices are provided to specify even a single primitive or when an incorrect multiple of vertices is specified. The incomplete primitive is ignored; the rest are drawn.

The minimum specification of vertices for each primitive is as follows: 1 for a point, 2 for a line, 3 for a triangle, 4 for a quadrilateral, and 3 for a polygon. Modes that require a certain multiple of vertices are Lines (2), Triangles (3), Quads (4), and QuadStrip (2).

unsafeRenderPrimitive :: PrimitiveMode -> IO a -> IO a

A more efficient, but potentially dangerous version of renderPrimitive: The given action is not allowed to throw an exception.

primitiveRestart :: IO ()

Polygon Edges

data EdgeFlag

A vertex can begin an edge which lies in the interior of its polygon or on the polygon's boundary.

Constructors

BeginsInteriorEdge

BeginsBoundaryEdge

edgeFlag :: StateVar EdgeFlag

Each vertex of a polygon, separate triangle, or separate quadrilateral specified during renderPrimitive is marked as the start of either a boundary or nonboundary (interior) edge.

The vertices of connected triangles and connected quadrilaterals are always marked as boundary, regardless of the value of the edge flag.

Boundary and nonboundary edge flags on vertices are significant only if polygonMode is set to Point or Line.

Note that the current edge flag can be updated at any time, in particular during renderPrimitive.