Graphics.Rendering.OpenGL.GL.ObjectName

Description

Object names are explicitly handled identifiers for API objects, e.g. a texture object name in OpenGL or a buffer object name in OpenAL.

Documentation

class ObjectName a where

An ObjectName is an explicitly handled identifier for API objects, e.g. a texture object name in OpenGL or a buffer object name in OpenAL.

Methods

genObjectNames :: Int -> IO [a]

Generate a given number of object names, which are guaranteed to be unused. By generating the names, they become used.

deleteObjectNames :: [a] -> IO ()

Make the given object names available again, declaring them as unused.

isObjectName :: a -> IO Bool

Test if the given object name is currently in use, i.e. test if it has been generated, but not been deleted so far.