Graphics.Rendering.OpenGL.GL.QueryObjects

Description

This module corresponds to section 4.2 (Query Objects and Asynchronous Queries) of the OpenGL 4.4 specs.

Creating and Delimiting Queries

data QueryObject

type QueryIndex = GLuint

maxVertexStreams :: GettableStateVar QueryIndex

data QueryTarget

Constructors

SamplesPassed

AnySamplesPassed

AnySamplesPassedConservative

TimeElapsed

PrimitivesGenerated QueryIndex

TransformFeedbackPrimitivesWritten QueryIndex

beginQuery :: QueryTarget -> QueryObject -> IO ()

endQuery :: QueryTarget -> IO ()

withQuery :: QueryTarget -> QueryObject -> IO a -> IO a

Convenience function for an exception-safe combination of beginQuery and endQuery.

Query Target Queries

currentQuery :: QueryTarget -> GettableStateVar (Maybe QueryObject)

queryCounterBits :: QueryTarget -> GettableStateVar GLsizei

Query Object Queries

queryResultAvailable :: QueryObject -> GettableStateVar Bool

class Storable a => QueryResult a

Minimal complete definition

getQueryObjectv

queryResult :: QueryResult a => QueryObject -> GettableStateVar a

Time Queries

timestampQuery :: QueryObject -> IO ()

Record the time after all previous commands on the GL client and server state and the framebuffer have been fully realized

timestamp :: GettableStateVar GLuint64

Contains the GL time after all previous commands have reached the GL server but have not yet necessarily executed.