Graphics.Rendering.OpenGL.GL.StringQueries

Description

This module corresponds to parts of section 6.1.5 (String Queries) of the OpenGL 3.2 specs.

Documentation

vendor :: GettableStateVar String

renderer :: GettableStateVar String

glVersion :: GettableStateVar String

glExtensions :: GettableStateVar [String]

shadingLanguageVersion :: GettableStateVar String

majorMinor :: GettableStateVar String -> GettableStateVar (Int, Int)

A utility function to be used with e.g. glVersion or shadingLanguageVersion, transforming a variable containing a string of the form major.minor[optional rest] into a variable containing a numeric major/minor version. If the string is malformed, which should never happen with a sane OpenGL implementation, it is transformed to (-1,-1).

contextProfile :: GettableStateVar [ContextProfile']