Graphics.Rendering.OpenGL.GL.TransformFeedback

Description

starting and ending

beginTransformFeedback :: PrimitiveMode -> IO ()

endTransformFeedback :: IO ()

TransformFeedbackBufferMode

data TransformFeedbackBufferMode

Constructors

InterleavedAttribs

SeperateAttribs

marshalTransformFeedbackBufferMode :: TransformFeedbackBufferMode -> GLenum

unmarshalTransformFeedbackBufferMode :: GLenum -> TransformFeedbackBufferMode

Shader related

transformFeedbackBufferMode :: Program -> GettableStateVar TransformFeedbackBufferMode

Get the currently used transformFeedbackBufferMode

getTransformFeedbackVaryings :: Program -> GettableStateVar GLuint

The number of varyings that are currently recorded when in transform feedback mode

setTransformFeedbackVaryings :: Program -> [String] -> TransformFeedbackBufferMode -> IO ()

Set all the transform feedbacks varyings for this program it overwrites any previous call to this function

getTransformFeedbackVarying

Arguments

:: Program

-> VaryingIndex
the index in a previous array of names of setTransformFeedbackVaryings

-> MaxLength
the maximum length of the returned string

-> IO (String, DataType, GLsizei)
The name of the varying, it's type and size

Get the name, datatype and size of a single transform feedback varying.

getTransformFeedbackVaryingMaxLength :: Program -> GettableStateVar GLuint

The maximum length of a varying's name for transform feedback mode

limits

maxTransformFeedbackSeparateAttribs :: GettableStateVar GLint

Max number of seprate atributes or varyings than can be captured in transformfeedback, initial value 4

maxTransformFeedbackInterleavedComponents :: GettableStateVar GLint

Max number of components to write to a single buffer in interleaved mod, initial value 64

maxTransformFeedbackSeparateComponents :: GettableStateVar GLint

Max number of components per attribute or varying in seperate mode initial value 4