Graphics.Rendering.OpenGL.GL.TransformFeedback

Description

starting and ending

beginTransformFeedback :: PrimitiveMode -> IO ()

endTransformFeedback :: IO ()

TransformFeedbackBufferMode

data TransformFeedbackBufferMode

Constructors

InterleavedAttribs

SeperateAttribs

marshalTransformFeedbackBufferMode :: TransformFeedbackBufferMode -> GLenum

unmarshalTransformFeedbackBufferMode :: GLenum -> TransformFeedbackBufferMode

Shader related

transformFeedbackBufferMode :: Program -> GettableStateVar TransformFeedbackBufferMode

Get the currently used transformFeedbackBufferMode

transformFeedbackVaryings :: Program -> GettableStateVar [(GLint, DataType, String)]

The name, datatype and size of the transform feedback varyings.

setTransformFeedbackVaryings :: Program -> [String] -> TransformFeedbackBufferMode -> IO ()

Set all the transform feedbacks varyings for this program it overwrites any previous call to this function

limits

maxTransformFeedbackSeparateAttribs :: GettableStateVar GLint

Max number of seprate atributes or varyings than can be captured in transformfeedback, initial value 4

maxTransformFeedbackInterleavedComponents :: GettableStateVar GLint

Max number of components to write to a single buffer in interleaved mod, initial value 64

maxTransformFeedbackSeparateComponents :: GettableStateVar GLint

Max number of components per attribute or varying in seperate mode initial value 4