Graphics.Rendering.OpenGL.GLU.Errors

Description

This module corresponds to section 2.5 (GL Errors) of the OpenGL 2.1 specs and chapter 8 (Errors) of the GLU specs, offering a generalized view of errors in GL and GLU.

Documentation

data Error

GL/GLU errors consist of a general error category and a description of what went wrong.

Constructors

Error ErrorCategory String

data ErrorCategory

General GL/GLU error categories

Constructors

InvalidEnum

InvalidValue

InvalidOperation

InvalidFramebufferOperation

OutOfMemory

StackOverflow

StackUnderflow

TableTooLarge

TesselatorError

NURBSError

errors :: GettableStateVar [Error]

When an error occurs, it is recorded in this state variable and no further errors are recorded. Reading errors returns the currently recorded errors (there may be more than one due to a possibly distributed implementation) and resets the state variable to [], re-enabling the recording of future errors. The value [] means that there has been no detectable error since the last time errors was read, or since the GL was initialized.